Address comments from James
* Refactor integration tests, move class set up to PublishAppIntegBase
* Use UUID as application name in integration tests
* Remove S3 permission exception wrapping (moved to serverlessrepo python library)
* Add exception handling and tests for publishing not packaged template
