+++ tests/integration/testdata/publish/metadata_create_app.json
@@ -6,7 +6,7 @@
     "test-app"
   ],
   "LicenseUrl": "s3://<bucket-name>/LICENSE",
-  "Name": "hello-world",
+  "Name": "<application-name>",
   "ReadmeUrl": "s3://<bucket-name>/README.rst",
   "SemanticVersion": "0.0.1",
   "SourceCodeUrl": "https://github.com/test/test"
