+++ tests/integration/testdata/publish/s3_bucket_policy.json
@@ -8,14 +8,6 @@
             },
             "Action": "s3:GetObject",
             "Resource": "arn:aws:s3:::<bucket-name>/*"
-        },
-        {
-            "Effect": "Deny",
-            "Principal": {
-                "Service": "serverlessrepo.amazonaws.com"
-            },
-            "Action": "s3:GetObject",
-            "Resource": "arn:aws:s3:::<bucket-name>/README_NO_PERMISSION.rst"
         }
     ]
 }
