+++ tests/integration/testdata/publish/template_update_app.yaml
@@ -15,7 +15,7 @@ Metadata:
     Labels:
     - test-app-update
     LicenseUrl: s3://<bucket-name>/LICENSE
-    Name: hello-world
+    Name: <application-name>
     ReadmeUrl: s3://<bucket-name>/README_UPDATE.rst
     SemanticVersion: 0.0.1
     SourceCodeUrl: https://github.com/test/test
